+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 /**
  * main - Entry point
  *
@@ -7,12 +8,12 @@
 
 int main(void)
 {
-	unsigned long long int i, fib1 = 1, fib2 = 2;
+	uintmax_t i, fib1 = 1, fib2 = 2;
 
 	for (i = 0; i < 99; i++)
 	{
-		printf("%llu, ", fib1);
-		printf("%llu, ", fib2);
+		printf("%ju, ", fib1);
+		printf("%ju, ", fib2);
 		fib1 = fib1 + fib2;
 		fib2 = fib1 + fib2;
 	}
